resolves bug c7
