+++ central_v2/src/lib/context/CentralDataContext.tsx
@@ -1,65 +1,6 @@
 import React, { createContext, useReducer } from 'react';
-import { SortDirection, SortType, PublicPrivateType } from '@righton/networking';
 import { centralDataReducer, CentralDataAction } from '../reducer/CentralDataReducer';
-import { ICentralDataState } from './ICentralDataState';
-import { UserStatusType } from '../CentralModels';
-
-export const userProfileInit = {
-    title: 'Title...',
-    firstName: '',
-    lastName: '',
-    userName: '',
-    email: '',
-    password: '',
-    gamesUsed: 0,
-}
-
-// initialize centralDataState
-const initCentralDataState: ICentralDataState = {
-  userProfile: userProfileInit,
-  userStatus: UserStatusType.LOADING,
-  recommendedGames: [],
-  mostPopularGames: [],
-  searchedGames: [],
-  publicGames: [],
-  privateGames: [],
-  draftGames: [],
-  favGames: [],
-  selectedGame: {
-    game: null,
-    profilePic: '',
-    createdName: '',
-    lastModified: new Date(),
-    timesPlayed: 0
-  },
-  recommendedQuestions: [],
-  mostPopularQuestions: [],
-  searchedQuestions: [],
-  publicQuestions: [],
-  privateQuestions: [],
-  draftQuestions: [],
-  favQuestions: [],
-  selectedQuestion: {
-    question: null,
-    profilePic: '',
-    createdName: '',
-    lastModified: new Date(),
-    timesPlayed: 0
-  },
-  nextToken: null,
-  isLoading: false,
-  isLoadingInfiniteScroll: false,
-  searchTerms: '',
-  selectedGrades: [],
-  isLibraryInit: true,
-  isTabsOpen: false,
-  isFavTabOpen: false,
-  publicPrivate: PublicPrivateType.PUBLIC,
-  sort: {
-    field: SortType.listGameTemplates,
-    direction: SortDirection.ASC,
-  }
-}
+import { ICentralDataState, initCentralDataState } from '../CentralModels';
 
 export interface CentralDataContextType {
   centralData: ICentralDataState;
